Add Discord Notifications to Backup Script
In this commit, we enhanced the Docker backup script by integrating it with Discord notifications. The notifications are sent through a Discord webhook which enables real-time updates about the backup process.

Here are the key changes made:

Discord Webhook Integration: Added a Discord webhook at the end of the backup process for each Docker container. This webhook sends a message to the specified Discord channel when a backup has been completed.

Dynamic Messages: The Discord messages are dynamically generated to include the name of the Docker container (ARCHIVE), the folder where the backup is stored (STORAGE), and the current timestamp when the backup was completed.

Improved Tracking: With these notifications, users can monitor the progress of the backup process in real-time and can immediately know if a backup has been successfully completed or if there are issues that need to be addressed.

This feature adds a new layer of visibility to the backup process and makes it easier for users to manage their Docker containers.

Please ensure to replace the placeholder "[Your Discord Webhook URL Here]" with your actual Discord webhook URL to make use of this feature.

Remember to keep this script secure as it contains sensitive information like your Discord Webhook URL. The script must not be shared publicly or with untrusted parties to prevent misuse.

The changes are designed to improve user experience and provide an efficient way

to keep track of Docker backups. As always, we appreciate any feedback or suggestions for further improvements.

Signed-off-by: DoCtEuR3805 | FRENCH-QC <35934387+dan3805@users.noreply.github.com>